Move OpenSSL specific tlsver.hpp code to OpenSSL specific tlsverp.hpp
Replace the #ifdef OPENVPN_SSL_TLSVER_H that suggest the file is OpenSSL
specific with a pragma once
